requirements.txt
# Core Django packages
Django==5.2
djangorestframework==3.16.0
django-cors-headers==4.3.0
django-filter==23.3
djangorestframework-simplejwt==5.3.0

asgiref==3.8.1
sqlparse==0.5.3
# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Caching
redis==5.0.1
django-redis==5.4.0

# Task Queue
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1

# API Documentation
drf-yasg==1.21.7

# Environment variables
python-dotenv==1.0.0

# Storage
django-storages==1.14.2
boto3==1.28.64

# Utilities
Pillow==10.1.0
pandas==2.1.2
numpy==1.26.1

# Testing
pytest==7.4.3
pytest-django==4.5.2
factory-boy==3.3.0

# Development
black==23.10.1
flake8==6.1.0
isort==5.12.0
